Reformat project structure
