+++ src/components/Board/index.js
@@ -1,0 +1,3 @@
+import { Board } from './styles';
+
+export default Board;
