+++ src/components/Keycap/index.js
@@ -1,0 +1,3 @@
+import { Keycap } from './styles';
+
+export default Keycap;
