+++ src/components/Keyrow/index.js
@@ -1,0 +1,3 @@
+import { Keyrow } from './styles';
+
+export default Keyrow;
