+++ src/components/Keyrow/styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+import { KeyrowUnstyled } from './Keyrow';
+
+export const Keyrow = styled(KeyrowUnstyled)`
+    display: flex;
+    align-items: center;
+    justify-content: space-around;
+`;
