+++ src/components/LetterBox/index.js
@@ -1,0 +1,3 @@
+import { LetterBox } from './styles';
+
+export default LetterBox;
