+++ src/components/OnScreenKeyboard/index.js
@@ -1,0 +1,3 @@
+import { OnScreenKeyboard } from './styles';
+
+export default OnScreenKeyboard;
