+++ src/components/OnScreenKeyboard/styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+import { OnScreenKeyboardUnstyled } from './OnScreenKeyboard';
+
+
+export const OnScreenKeyboard = styled(OnScreenKeyboardUnstyled)`
+    display: table;
+    color: aliceblue;
+`;
