+++ src/utils/Utils.js
@@ -1,3 +1,9 @@
+
+export const zip = (...arrays) => {
+    const length = Math.min(...arrays.map(arr => arr.length));
+    return Array.from({ length }, (value, index) => arrays.map((array => array[index])));
+};
+
 export function checkWord(guess, solution) {
 
     //TODO: if allowed words don't contain guess -> invalid
